UnitsList: highlight selected unit in list
This is a hackish custom implementation of single-select, (ab)using the
fact that as of Leptonic 0.5 the <TableRow> component does nothing more than
generating a <leptonic-table-row>.  We cannot pass an Leptos dynamic class
to a component, so just create said element directly without going through
a component, and set the dynamic class on it.

Hopefully we'll be able to do better in the future.