Switch to using leptos_router, but not for routes
Routes as currently implemented will not work with a dumb http server,
so we won't use them when it comes to a separate view for army selection,
but this will allow this view to access URL parameters for armies to be
used, using Leptos reactive mechanism.

Router component must also contain all other components that need to get
the query string, so lives in AppBoilerplate.