Replace cascaded set_/shown signal with a context shared with ArmyList
This will allow to hide the drawer when army changes.

Note the use of <Provider> instead of a simple provide_context, to
isolate each army container in a separate reactive scope - otherwise
there the context would only be set properly for the component tree,
and the dynamically-created ones would only see a single context for
all armies (effectively causing all drawers' "close" button to close
a single drawer).